+++ admin/personnel/index.php
@@ -7,7 +7,7 @@ if (!isset($_SESSION["nom_utilisateur"])) {
     exit();
 }
 
-if ($_SESSION["role"] != "Directeur") {
+if ($_SESSION["role"] != "Directeur" && $_SESSION["role"] != "CM" && $_SESSION["role"] != "Responsable") {
     header("Location: ../index.php");
     exit();
 }
@@ -60,25 +60,37 @@ Include("role.php") ;
 </div>
 <h1 class="title">Liste Personnel</h1>
 <br>
-<a href="ajouter.php" style="display: inline-block; padding: 10px 20px; background-color: #ffffff; border: 1px solid #000000;">Ajouter un Personnel</a>
+<?php if ($_SESSION["role"] == "Directeur"){
+    echo '<a href="ajouter.php" style="display: inline-block; padding: 10px 20px; background-color: #ffffff; border: 1px solid #000000;">Ajouter un Personnel</a>';
+}?>
+
 <div class="resultat">
 
     <?php
     $sql = "SELECT * FROM Personnel";
     $result = $conn->query($sql);
+    if ($_SESSION["role"] == "Directeur") {
+        echo "<table><thead><tr><th>Numéro de sécurité sociale</th><th>Nom</th><th>Prenom</th><th>Role</th><th>Modifier</th><th>Supprimer</th></tr><thead><tbody>";
+        while($row = $result->fetch_assoc()) {
+            echo "<tr><td>" . $row["Numero_SS"]. "</td><td>" . $row["Nom"]. "</td><td>" . $row["Prenom"]. "</td>";
+            echo "<td>" . chercherRole($row["Numero_SS"]) . "</td>";
+            if ($row["Numero_SS"] != $_SESSION["numero_ss"]) {
+                echo "<td><a href='modifier.php?id=" . $row["Numero_SS"] . "'>Modifier</a></td>";
+                echo "<td><a href='supprimer.php?id=" . $row["Numero_SS"] . "'>Supprimer</a></td>";
+            } else {
+                echo "<td>bloqué</td><td>bloqué</td>";
+            }
+        }
+    } else {
+        echo "<table><thead><tr><th>Numéro de sécurité sociale</th><th>Nom</th><th>Prenom</th><th>Role</th></tr><thead><tbody>";
+        while($row = $result->fetch_assoc()) {
+            echo "<tr><td>" . $row["Numero_SS"]. "</td><td>" . $row["Nom"]. "</td><td>" . $row["Prenom"]. "</td>";
+            echo "<td>" . chercherRole($row["Numero_SS"]) . "</td>";
 
-    // Affichage des manèges dans un tableau
-    echo "<table><thead><tr><th>Numéro de sécurité sociale</th><th>Nom</th><th>Prenom</th><th>Role</th>><th>Modifier</th><th>Supprimer</th></tr><thead><tbody>";
-    while($row = $result->fetch_assoc()) {
-        echo "<tr><td>" . $row["Numero_SS"]. "</td><td>" . $row["Nom"]. "</td><td>" . $row["Prenom"]. "</td>";
-        echo "<td>" . chercherRole($row["Numero_SS"]) . "</td>";
-        if ($row["Numero_SS"] != $_SESSION["numero_ss"]) {
-            echo "<td><a href='modifier.php?id=" . $row["Numero_SS"] . "'>Modifier</a></td>";
-            echo "<td><a href='supprimer.php?id=" . $row["Numero_SS"] . "'>Supprimer</a></td>";
-        } else {
-            echo "<td>bloqué</td><td>bloqué</td>";
         }
     }
+    // Affichage des manèges dans un tableau
+
     echo "</tbody></table>";
 
     // Fermeture de la connexion à la base de données
